refactor: implement phase 2 simplified routing architecture
- Update EnhancedNeuralAdapter to be primary implementation
- Remove mock adapter references from all modules
- Implement performance channel infrastructure
- Update DAA coordinator for proper initialization
- Refactor config module for better organization
- Update examples to use new architecture

Work in progress: ~70 compilation errors remain to be fixed

+++ src/neural/tests/test_performance_regression.rs
@@ -307,7 +307,7 @@ fn get_process_memory_mb() -> f64 {
 
     let pid = std::process::id();
     system
-        .process(pid.into())
+        .process((pid as usize).into())
         .map(|p| p.memory() as f64 / 1024.0 / 1024.0)
         .unwrap_or(0.0)
 }
